Fix up casting status to int8_t.

+++ src/protocols/ab/ab_common.c
@@ -208,7 +208,7 @@ plc_tag_p ab_tag_create(attr attribs)
 
     if(check_cpu(tag, attribs) != PLCTAG_STATUS_OK) {
         pdebug(DEBUG_WARN,"CPU type not valid or missing.");
-        /* tag->status = PLCTAG_ERR_BAD_DEVICE; */
+        /* tag->status = (int8_t)PLCTAG_ERR_BAD_DEVICE; */
         rc_dec(tag);
         return (plc_tag_p)NULL;
     }
@@ -262,7 +262,7 @@ plc_tag_p ab_tag_create(attr attribs)
 
     default:
         pdebug(DEBUG_WARN, "Unknown PLC type!");
-        tag->status = PLCTAG_ERR_BAD_CONFIG;
+        tag->status = (int8_t)PLCTAG_ERR_BAD_CONFIG;
         return (plc_tag_p)tag;
         break;
     }
@@ -280,7 +280,7 @@ plc_tag_p ab_tag_create(attr attribs)
      */
     if(session_find_or_create(&tag->session, attribs) != PLCTAG_STATUS_OK) {
         pdebug(DEBUG_INFO,"Unable to create session!");
-        tag->status = PLCTAG_ERR_BAD_GATEWAY;
+        tag->status = (int8_t)PLCTAG_ERR_BAD_GATEWAY;
         return (plc_tag_p)tag;
     }
 
@@ -355,7 +355,7 @@ plc_tag_p ab_tag_create(attr attribs)
         /* Logix tags need a path. */
         if(path == NULL && tag->plc_type == AB_PLC_LGX) {
             pdebug(DEBUG_WARN,"A path is required for Logix-class PLCs!");
-            tag->status = PLCTAG_ERR_BAD_PARAM;
+            tag->status = (int8_t)PLCTAG_ERR_BAD_PARAM;
             return (plc_tag_p)tag;
         }
 
@@ -406,7 +406,7 @@ plc_tag_p ab_tag_create(attr attribs)
 
         if(str_length(path) == 0) {
             pdebug(DEBUG_WARN,"A path is required for this PLC type.");
-            tag->status = PLCTAG_ERR_BAD_PARAM;
+            tag->status = (int8_t)PLCTAG_ERR_BAD_PARAM;
             return (plc_tag_p)tag;
         }
 
@@ -429,7 +429,7 @@ plc_tag_p ab_tag_create(attr attribs)
 
     default:
         pdebug(DEBUG_WARN, "Unknown PLC type!");
-        tag->status = PLCTAG_ERR_BAD_CONFIG;
+        tag->status = (int8_t)PLCTAG_ERR_BAD_CONFIG;
         return (plc_tag_p)tag;
     }
 
@@ -469,7 +469,7 @@ plc_tag_p ab_tag_create(attr attribs)
         if(tag->size == 0) {
             /* failure! Need data_size! */
             pdebug(DEBUG_WARN,"Tag size is zero!");
-            tag->status = PLCTAG_ERR_BAD_PARAM;
+            tag->status = (int8_t)PLCTAG_ERR_BAD_PARAM;
             return (plc_tag_p)tag;
         }
 
@@ -478,7 +478,7 @@ plc_tag_p ab_tag_create(attr attribs)
 
         if(tag->data == NULL) {
             pdebug(DEBUG_WARN,"Unable to allocate tag data!");
-            tag->status = PLCTAG_ERR_NO_MEM;
+            tag->status = (int8_t)PLCTAG_ERR_NO_MEM;
             return (plc_tag_p)tag;
         }
         break;
@@ -490,7 +490,7 @@ plc_tag_p ab_tag_create(attr attribs)
 
     if(!tag->special_tag && check_tag_name(tag, attr_get_str(attribs,"name",NULL)) != PLCTAG_STATUS_OK) {
         pdebug(DEBUG_INFO,"Bad tag name!");
-        tag->status = PLCTAG_ERR_BAD_PARAM;
+        tag->status = (int8_t)PLCTAG_ERR_BAD_PARAM;
         return (plc_tag_p)tag;
     }
 
@@ -855,7 +855,7 @@ int ab_get_int_attrib(plc_tag_p raw_tag, const char *attrib_name, int default_va
     pdebug(DEBUG_SPEW, "Starting.");
 
     /* assume we have a match. */
-    tag->status = PLCTAG_STATUS_OK;
+    tag->status = (int8_t)PLCTAG_STATUS_OK;
 
     /* match the attribute. */
     if(str_cmp_i(attrib_name, "elem_size") == 0) {
@@ -864,7 +864,7 @@ int ab_get_int_attrib(plc_tag_p raw_tag, const char *attrib_name, int default_va
         res = tag->elem_count;
     } else {
         pdebug(DEBUG_WARN, "Unsupported attribute name \"%s\"!", attrib_name);
-        tag->status = PLCTAG_ERR_UNSUPPORTED;
+        tag->status = (int8_t)PLCTAG_ERR_UNSUPPORTED;
     }
 
     return res;
